Slight changes to scripts to make it repeatable

+++ paas-node/packer/scripts/install.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+#Check if docker already installed and exit
+DOCKER_EXISTS=`dpkg -l | grep lxc-docker`
+if [ "$DOCKER_EXISTS" != "" ]
+then
+	exit 0
+fi
+
 sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys 36A1D7869245C8950F966E92D8576A8BA88D21E9
 echo "deb https://get.docker.com/ubuntu docker main" > /etc/apt/sources.list.d/docker.list
